EVO | Features routings global

+++ src/main/java/wishlist/service/UserService.java
@@ -5,14 +5,20 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import wishlist.Utils;
+import wishlist.dto.GiftDTO;
 import wishlist.dto.GroupDTO;
 import wishlist.dto.UserDTO;
+import wishlist.entity.Gift;
 import wishlist.entity.Group;
-import wishlist.entity.GroupMember;
+import wishlist.entity.Reservation;
 import wishlist.entity.User;
+import wishlist.mapper.GiftMapper;
 import wishlist.mapper.GroupMapper;
 import wishlist.mapper.UserMapper;
+import wishlist.repository.GiftRepository;
 import wishlist.repository.GroupMembersRepository;
+import wishlist.repository.ReservationsRepository;
 import wishlist.repository.UserRepository;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
@@ -24,8 +30,11 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class UserService implements UserDetailsService {
 
+    private final GiftRepository giftRepository;
+    private final GiftMapper giftMapper;
     private final GroupMembersRepository groupMembersRepository;
     private final UserRepository userRepository;
+    private final ReservationsRepository reservationsRepository;
     private final UserMapper userMapper;
     private final GroupMapper groupMapper;
 
@@ -60,16 +69,39 @@ public class UserService implements UserDetailsService {
                 .collect(Collectors.toList());
     }
 
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public List<GiftDTO> getAllGiftOfUserForGroup(Long id,Long groupID) {
+        List<Gift> giftDTOList = giftRepository.findGiftsByUserIdAndGroupId(id, groupID);
+        return giftDTOList.stream()
+                .map(giftMapper::toDTO)
+                .collect(Collectors.toList());
     }
 
     public UserDTO updateUser(Long id, UserDTO userDTO) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("User not found"));
-        user.setUsername(userDTO.getUsername());
-        user.setRole(userDTO.getRole());
-        user.setEmail(userDTO.getEmail());
-        return userMapper.toDTO(userRepository.save(user));
+
+        Utils.updateEntityFromDTO(userDTO, user);
+
+        User updatedUser = userRepository.save(user);
+        return userMapper.toDTO(updatedUser);
+    }
+
+    public boolean reserveGift(Long userId, Long giftId) {
+        if (userRepository.existsById(userId) && giftRepository.existsById(giftId)) {
+            Reservation reservation = new Reservation();
+            reservation.setReserved_by(userId);
+            reservation.setGiftId(giftId);
+            reservationsRepository.save(reservation);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean deleteUser(Long id) {
+        if(userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            return !userRepository.existsById(id);
+        }
+        return false;
     }
 }
